# GODRIVER-2677 Limit the maximum number of items in pool.