--- a/x/mongo/driver/operation_exhaust.go
+++ b/x/mongo/driver/operation_exhaust.go
@@ -18,13 +18,15 @@ func (op Operation) ExecuteExhaust(ctx context.Context, conn StreamerConnection)
 		return errors.New("exhaust read must be done with a connection that is currently streaming")
 	}
 
-	wm := memoryPool.Get().(*[]byte)
+	wm, err := memoryPool.Get(context.Background())
+	if err != nil {
+		return errors.New("failed to allocate memory")
+	}
 	defer func() {
 		memoryPool.Put(wm)
 	}()
 	var res []byte
-	var err error
-	res, *wm, err = op.readWireMessage(ctx, conn, (*wm)[:0])
+	res, wm, err = op.readWireMessage(ctx, conn, wm[:0])
 	if err != nil {
 		return err
 	}