GODRIVER-2677

## Summary
Use a semaphore to limit the maximum number of items in the byte slice pool.

## Background & Motivation
I limited 512 slices in the pool. `Get()` blocks until a slice is returned to the pool when all are allocated. So the size of the pool shall not be bigger than 16MB * 512 ~= 8GB in theory considering the ticket indicates ~20GB of memory consumption.
I did not reset the capacity of returned byte slice because it would cause more allocations.

The benchstat from `benchmark/operation_test.go`:
```
name                           old time/op    new time/op    delta
ClientWrite/not_compressed-10    2.17ms ±16%    1.94ms ± 4%  -10.45%  (p=0.005 n=10+10)
ClientWrite/snappy-10            2.04ms ±18%    1.96ms ±10%     ~     (p=0.631 n=10+10)
ClientWrite/zlib-10              1.85ms ±12%    1.93ms ±11%     ~     (p=0.218 n=10+10)
ClientWrite/zstd-10              1.96ms ±10%    1.99ms ± 8%     ~     (p=0.684 n=10+10)

name                           old alloc/op   new alloc/op   delta
ClientWrite/not_compressed-10    27.9kB ± 2%    27.6kB ± 2%   -1.16%  (p=0.035 n=10+10)
ClientWrite/snappy-10            41.5kB ± 1%    41.2kB ± 1%     ~     (p=0.075 n=10+10)
ClientWrite/zlib-10               906kB ± 0%     907kB ± 0%   +0.11%  (p=0.000 n=9+9)
ClientWrite/zstd-10              67.0kB ± 0%    67.1kB ± 0%     ~     (p=0.165 n=10+10)

name                           old allocs/op  new allocs/op  delta
ClientWrite/not_compressed-10      79.9 ± 1%      81.0 ± 0%   +1.38%  (p=0.001 n=10+10)
ClientWrite/snappy-10              82.3 ± 2%      85.0 ± 0%   +3.28%  (p=0.000 n=10+9)
ClientWrite/zlib-10                 122 ± 0%       126 ± 0%   +2.95%  (p=0.000 n=6+10)
ClientWrite/zstd-10                 183 ± 0%       186 ± 1%   +1.58%  (p=0.000 n=9+10)
```

```
name                          old time/op    new time/op    delta
ClientRead/not_compressed-10     140µs ± 2%     153µs ± 5%   +9.15%  (p=0.000 n=9+9)
ClientRead/snappy-10             119µs ± 1%     170µs ±22%  +42.64%  (p=0.000 n=9+10)
ClientRead/zlib-10               218µs ± 6%     257µs ± 3%  +17.82%  (p=0.000 n=10+9)
ClientRead/zstd-10               117µs ± 1%     156µs ± 1%  +32.95%  (p=0.000 n=9+8)

name                          old alloc/op   new alloc/op   delta
ClientRead/not_compressed-10    29.0kB ± 0%    29.0kB ± 0%     ~     (p=0.753 n=10+10)
ClientRead/snappy-10            45.8kB ± 0%    45.7kB ± 0%   -0.10%  (p=0.012 n=10+10)
ClientRead/zlib-10               905kB ± 0%     905kB ± 0%   -0.02%  (p=0.009 n=10+10)
ClientRead/zstd-10               135kB ± 0%     136kB ± 0%   +0.08%  (p=0.000 n=10+10)

name                          old allocs/op  new allocs/op  delta
ClientRead/not_compressed-10      83.0 ± 0%      85.0 ± 0%   +2.41%  (p=0.000 n=10+10)
ClientRead/snappy-10              87.0 ± 0%      90.0 ± 0%   +3.45%  (p=0.000 n=10+10)
ClientRead/zlib-10                 143 ± 0%       146 ± 0%   +2.10%  (p=0.000 n=7+9)
ClientRead/zstd-10                 191 ± 0%       194 ± 0%   +1.57%  (p=0.000 n=10+10)
```

The number of allocations increases, but it does not impact the size of allocations.
For the increase of reading time, my understanding is the overhead operation, which is relatively big compared with its base (reading time is in µs while writing time is in ms).